Refactor package generators
